// generated/api-client/models/blog-summary-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BlogSummaryResponseDtoAuthor } from './blog-summary-response-dto-author';
// May contain unused imports in some cases
// @ts-ignore
import { MyBlogResponseDtoCoverImagesInner } from './my-blog-response-dto-cover-images-inner';
// May contain unused imports in some cases
// @ts-ignore
import { MyBlogResponseDtoTagsInnerTag } from './my-blog-response-dto-tags-inner-tag';

/**
 * 
 * @export
 * @interface BlogSummaryResponseDto
 */
export interface BlogSummaryResponseDto {
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'subtitle': string | null;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'excerpt': string | null;
    /**
     * 
     * @type {number}
     * @memberof BlogSummaryResponseDto
     */
    'readingTime': number | null;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'publishedAt': string | null;
    /**
     * 
     * @type {number}
     * @memberof BlogSummaryResponseDto
     */
    'viewCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogSummaryResponseDto
     */
    'likeCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogSummaryResponseDto
     */
    'commentCount': number;
    /**
     * 
     * @type {number}
     * @memberof BlogSummaryResponseDto
     */
    'bookmarkCount': number;
    /**
     * 
     * @type {boolean}
     * @memberof BlogSummaryResponseDto
     */
    'isPinned': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogSummaryResponseDto
     */
    'isFeatured': boolean;
    /**
     * 
     * @type {BlogSummaryResponseDtoAuthor}
     * @memberof BlogSummaryResponseDto
     */
    'author': BlogSummaryResponseDtoAuthor;
    /**
     * 
     * @type {Array<MyBlogResponseDtoCoverImagesInner>}
     * @memberof BlogSummaryResponseDto
     */
    'coverImages': Array<MyBlogResponseDtoCoverImagesInner>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoTagsInnerTag>}
     * @memberof BlogSummaryResponseDto
     */
    'tags': Array<MyBlogResponseDtoTagsInnerTag>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoTagsInnerTag>}
     * @memberof BlogSummaryResponseDto
     */
    'categories': Array<MyBlogResponseDtoTagsInnerTag>;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof BlogSummaryResponseDto
     */
    'updatedAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof BlogSummaryResponseDto
     */
    'isLiked': boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof BlogSummaryResponseDto
     */
    'isBookmarked': boolean | null;
}