// generated/api-client/models/create-blog-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { CreateBlogDtoCoverImagesInner } from './create-blog-dto-cover-images-inner';

/**
 * 
 * @export
 * @interface CreateBlogDto
 */
export interface CreateBlogDto {
    /**
     * Blog title
     * @type {string}
     * @memberof CreateBlogDto
     */
    'title': string;
    /**
     * Blog subtitle
     * @type {string}
     * @memberof CreateBlogDto
     */
    'subtitle'?: string;
    /**
     * Blog content in markdown format
     * @type {string}
     * @memberof CreateBlogDto
     */
    'content': string;
    /**
     * Short excerpt for previews
     * @type {string}
     * @memberof CreateBlogDto
     */
    'excerpt'?: string;
    /**
     * Meta description for SEO
     * @type {string}
     * @memberof CreateBlogDto
     */
    'metaDescription'?: string;
    /**
     * Meta keywords for SEO
     * @type {Array<Array>}
     * @memberof CreateBlogDto
     */
    'metaKeywords'?: Array<Array>;
    /**
     * Canonical URL for republished content
     * @type {string}
     * @memberof CreateBlogDto
     */
    'canonicalUrl'?: string;
    /**
     * Blog tags
     * @type {Array<Array>}
     * @memberof CreateBlogDto
     */
    'tagNames'?: Array<Array>;
    /**
     * Blog categories
     * @type {Array<Array>}
     * @memberof CreateBlogDto
     */
    'categoryNames'?: Array<Array>;
    /**
     * Cover images for the blog
     * @type {Array<CreateBlogDtoCoverImagesInner>}
     * @memberof CreateBlogDto
     */
    'coverImages'?: Array<CreateBlogDtoCoverImagesInner>;
    /**
     * Allow comments on this blog
     * @type {boolean}
     * @memberof CreateBlogDto
     */
    'allowComments'?: boolean;
    /**
     * Allow bookmarks on this blog
     * @type {boolean}
     * @memberof CreateBlogDto
     */
    'allowBookmarks'?: boolean;
    /**
     * Allow likes on this blog
     * @type {boolean}
     * @memberof CreateBlogDto
     */
    'allowLikes'?: boolean;
    /**
     * Allow shares on this blog
     * @type {boolean}
     * @memberof CreateBlogDto
     */
    'allowShares'?: boolean;
    /**
     * Schedule the blog for future publishing
     * @type {string}
     * @memberof CreateBlogDto
     */
    'scheduledFor'?: string;
}