// generated/api-client/models/update-blog-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { CreateBlogDtoCoverImagesInner } from './create-blog-dto-cover-images-inner';

/**
 * 
 * @export
 * @interface UpdateBlogDto
 */
export interface UpdateBlogDto {
    /**
     * Blog title
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'title'?: string;
    /**
     * Blog subtitle
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'subtitle'?: string;
    /**
     * Blog content in markdown format
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'content'?: string;
    /**
     * Short excerpt for previews
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'excerpt'?: string;
    /**
     * Meta description for SEO
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'metaDescription'?: string;
    /**
     * Meta keywords for SEO
     * @type {Array<Array>}
     * @memberof UpdateBlogDto
     */
    'metaKeywords'?: Array<Array>;
    /**
     * Canonical URL for republished content
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'canonicalUrl'?: string;
    /**
     * Blog tags
     * @type {Array<Array>}
     * @memberof UpdateBlogDto
     */
    'tagNames'?: Array<Array>;
    /**
     * Blog categories
     * @type {Array<Array>}
     * @memberof UpdateBlogDto
     */
    'categoryNames'?: Array<Array>;
    /**
     * Cover images for the blog
     * @type {Array<CreateBlogDtoCoverImagesInner>}
     * @memberof UpdateBlogDto
     */
    'coverImages'?: Array<CreateBlogDtoCoverImagesInner>;
    /**
     * Allow comments on this blog
     * @type {boolean}
     * @memberof UpdateBlogDto
     */
    'allowComments'?: boolean;
    /**
     * Allow bookmarks on this blog
     * @type {boolean}
     * @memberof UpdateBlogDto
     */
    'allowBookmarks'?: boolean;
    /**
     * Allow likes on this blog
     * @type {boolean}
     * @memberof UpdateBlogDto
     */
    'allowLikes'?: boolean;
    /**
     * Allow shares on this blog
     * @type {boolean}
     * @memberof UpdateBlogDto
     */
    'allowShares'?: boolean;
    /**
     * Schedule the blog for future publishing
     * @type {string}
     * @memberof UpdateBlogDto
     */
    'scheduledFor'?: string;
}