// generated/api-client/models/flag-blog-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FlagBlogDto
 */
export interface FlagBlogDto {
    /**
     * Reason for flagging the blog
     * @type {string}
     * @memberof FlagBlogDto
     */
    'reason': FlagBlogDtoReasonEnum;
    /**
     * Additional description for the flag
     * @type {string}
     * @memberof FlagBlogDto
     */
    'description'?: string;
}

export const FlagBlogDtoReasonEnum = {
    SPAM: 'SPAM',
    HARASSMENT: 'HARASSMENT',
    HATE_SPEECH: 'HATE_SPEECH',
    INAPPROPRIATE_CONTENT: 'INAPPROPRIATE_CONTENT',
    COPYRIGHT_VIOLATION: 'COPYRIGHT_VIOLATION',
    MISINFORMATION: 'MISINFORMATION',
    VIOLENCE: 'VIOLENCE',
    ADULT_CONTENT: 'ADULT_CONTENT',
    OTHER: 'OTHER'
} as const;

export type FlagBlogDtoReasonEnum = typeof FlagBlogDtoReasonEnum[keyof typeof FlagBlogDtoReasonEnum];